requirements.txt
streamlit
ffmpeg-python
librosa
soundfile